chore(util): move check-valid-install out into separate util
no issue
- general code/test cleanup
- fix references to checkValidInstall everywhere
