+++ lib/command.js
@@ -4,9 +4,6 @@
  */
 'use strict';
 
-const fs = require('fs-extra');
-const path = require('path');
-const chalk = require('chalk');
 const each = require('lodash/each');
 const createDebug = require('debug');
 const kebabCase = require('lodash/kebabCase');
@@ -15,58 +12,6 @@ const UI = require('./ui');
 const System = require('./system');
 
 const debug = createDebug('ghost-cli:command');
-
-/**
- * Checks if the cwd is a valid ghost installation. Also checks if the current
- * directory contains a development install of ghost (e.g. a git clone), and outputs
- * a helpful error message if so.
- *
- * @param {string} name Name of command, will be output if there's an error
- */
-function checkValidInstall(name) {
-    /*
-     * CASE: a `config.js` file exists, which means this is a LTS installation
-     * which we don't support in the CLI
-     */
-    if (fs.existsSync(path.join(process.cwd(), 'config.js'))) {
-        console.error(
-            chalk.yellow('Ghost-CLI only works with Ghost versions >= 1.0.0.'),
-            '\nIf you are trying to upgrade Ghost LTS to 1.0.0 ' +
-            `please see ${chalk.blue.underline('https://docs.ghost.org/v1/docs/migrating-to-ghost-1-0-0')}.`,
-            `\nOtherwise, please run \`ghost ${name}\` again within a valid Ghost installation.`
-        );
-        process.exit(1);
-    }
-
-    /*
-     * We assume it's a Ghost development install if 3 things are true:
-     * 1) package.json exists
-     * 2) package.json "name" field is "ghost"
-     * 3) Gruntfile.js exists
-     */
-    if (fs.existsSync(path.join(process.cwd(), 'package.json')) &&
-        fs.readJsonSync(path.join(process.cwd(), 'package.json')).name === 'ghost' &&
-        fs.existsSync(path.join(process.cwd(), 'Gruntfile.js'))) {
-        console.error(
-            chalk.yellow('Ghost-CLI commands do not work inside of a git clone, zip download or with Ghost <1.0.0.'),
-            `\nPerhaps you meant \`grunt ${name}\`?`,
-            `\nOtherwise, please run \`ghost ${name}\` again within a valid Ghost installation.`
-        );
-        process.exit(1);
-    }
-
-    /*
-     * Assume it's not a valid CLI install if the `.ghost-cli` file doesn't exist
-     */
-    if (!fs.existsSync(path.join(process.cwd(), '.ghost-cli'))) {
-        console.error(
-            chalk.yellow('Working directory is not a recognisable Ghost installation.'),
-            `\nPlease run \`ghost ${name}\` again within a folder where Ghost was installed with Ghost-CLI.`
-        );
-
-        process.exit(1);
-    }
-}
 
 /**
  * Base Command class. All commands, both internal and external,
@@ -164,6 +109,8 @@ class Command {
      */
     static _run(commandName, argv, extensions) {
         if (!this.global) {
+            const checkValidInstall = require('./utils/check-valid-install');
+
             checkValidInstall(commandName);
         }
 
@@ -259,4 +206,3 @@ class Command {
 }
 
 module.exports = Command;
-module.exports.checkValidInstall = checkValidInstall;
