+++ lib/commands/doctor/checks/startup.js
@@ -7,7 +7,7 @@ const Promise   = require('bluebird');
 const errors = require('../../../errors');
 const Config = require('../../../utils/config');
 const advancedOptions = require('../../config/advanced');
-const checkValidInstall = require('../../../command').checkValidInstall;
+const checkValidInstall = require('../../../utils/check-valid-install');
 
 module.exports = [{
     title: 'Validating config',
