+++ lib/commands/log.js
@@ -8,11 +8,12 @@ class LogCommand extends Command {
         const includes = require('lodash/includes');
         const lastLines = require('read-last-lines');
         const PrettyStream = require('ghost-ignition/lib/logging/PrettyStream');
+        const checkValidInstall = require('../utils/check-valid-install');
 
         const errors = require('../errors');
 
         if (!argv.name) {
-            Command.checkValidInstall('log');
+            checkValidInstall('log');
         }
 
         const instance = this.system.getInstance(argv.name);
