+++ lib/commands/stop.js
@@ -21,6 +21,7 @@ class StopCommand extends Command {
     run(argv) {
         const errors = require('../errors');
         const ProcessManager = require('../process-manager');
+        const checkValidInstall = require('../utils/check-valid-install');
 
         const runOptions = {quiet: argv.quiet};
 
@@ -38,7 +39,7 @@ class StopCommand extends Command {
             process.chdir(instance.dir);
         }
 
-        Command.checkValidInstall('stop');
+        checkValidInstall('stop');
 
         if (!instance.running()) {
             return Promise.reject(new errors.SystemError('No running Ghost instance found here.'));
