+++ test/unit/command-spec.js
@@ -1,7 +1,6 @@
 'use strict';
 const expect = require('chai').expect;
 const sinon = require('sinon');
-const rewire = require('rewire');
 const proxyquire = require('proxyquire').noCallThru();
 
 const modulePath = '../../lib/command';
@@ -176,8 +175,9 @@ describe('Unit: Command', function () {
 
         it('calls checkValidInstall when global option is not set', function () {
             const checkValidInstall = sandbox.stub();
-            const Command = rewire(modulePath);
-            Command.__set__('checkValidInstall', checkValidInstall);
+            const Command = proxyquire(modulePath, {
+                './utils/check-valid-install': checkValidInstall
+            });
 
             const TestCommand = class extends Command {};
             checkValidInstall.throws();
@@ -440,113 +440,6 @@ describe('Unit: Command', function () {
                 expect(runSpy.calledOnce).to.be.true;
                 expect(runSpy.calledWithExactly({})).to.be.true;
             });
-        });
-    });
-
-    describe('checkValidInstall', function () {
-        const sandbox = sinon.sandbox.create();
-
-        afterEach(() => {
-            sandbox.restore();
-        })
-
-        it('throws error if config.js present', function () {
-            const existsStub = sandbox.stub();
-            existsStub.withArgs(sinon.match(/config\.js/)).returns(true);
-            const Command = proxyquire(modulePath, {
-                'fs-extra': {existsSync: existsStub}
-            });
-
-            const exitStub = sandbox.stub(process, 'exit').throws();
-            const errorStub = sandbox.stub(console, 'error');
-
-            try {
-                Command.checkValidInstall('test');
-                throw new Error('should not be thrown');
-            } catch (e) {
-                expect(e.message).to.not.equal('should not be thrown');
-                expect(existsStub.calledOnce).to.be.true;
-                expect(errorStub.calledOnce).to.be.true;
-                expect(exitStub.calledOnce).to.be.true;
-                expect(existsStub.args[0][0]).to.match(/config\.js/);
-                expect(errorStub.args[0][0]).to.match(/Ghost-CLI only works with Ghost versions >= 1\.0\.0/);
-            }
-        });
-
-        it('throws error if within a Ghost git clone', function () {
-            const existsStub = sandbox.stub();
-            const readJsonStub = sandbox.stub();
-
-            existsStub.withArgs(sinon.match(/config\.js/)).returns(false);
-            existsStub.withArgs(sinon.match(/package\.json/)).returns(true);
-            existsStub.withArgs(sinon.match(/Gruntfile\.js/)).returns(true);
-            readJsonStub.returns({name: 'ghost'});
-
-            const Command = proxyquire(modulePath, {
-                'fs-extra': {existsSync: existsStub, readJsonSync: readJsonStub}
-            });
-
-            const exitStub = sandbox.stub(process, 'exit').throws();
-            const errorStub = sandbox.stub(console, 'error');
-
-            try {
-                Command.checkValidInstall('test');
-                throw new Error('should not be thrown');
-            } catch (e) {
-                expect(e.message).to.not.equal('should not be thrown');
-                expect(existsStub.calledThrice).to.be.true;
-                expect(errorStub.calledOnce).to.be.true;
-                expect(exitStub.calledOnce).to.be.true;
-                expect(existsStub.args[1][0]).to.match(/package\.json/);
-                expect(errorStub.args[0][0]).to.match(/Ghost-CLI commands do not work inside of a git clone/);
-            }
-        });
-
-        it('throws error if above two conditions don\t exit and .ghost-cli file is missing', function () {
-            const existsStub = sandbox.stub();
-
-            existsStub.withArgs(sinon.match(/config\.js/)).returns(false);
-            existsStub.withArgs(sinon.match(/package\.json/)).returns(false);
-            existsStub.withArgs(sinon.match(/\.ghost-cli/)).returns(false);
-
-            const Command = proxyquire(modulePath, {
-                'fs-extra': {existsSync: existsStub}
-            });
-
-            const exitStub = sandbox.stub(process, 'exit').throws();
-            const errorStub = sandbox.stub(console, 'error');
-
-            try {
-                Command.checkValidInstall('test');
-                throw new Error('should not be thrown');
-            } catch (e) {
-                expect(e.message).to.not.equal('should not be thrown');
-                expect(existsStub.calledThrice).to.be.true;
-                expect(errorStub.calledOnce).to.be.true;
-                expect(exitStub.calledOnce).to.be.true;
-                expect(existsStub.args[2][0]).to.match(/\.ghost-cli/);
-                expect(errorStub.args[0][0]).to.match(/Working directory is not a recognisable Ghost installation/);
-            }
-        });
-
-        it('doesn\'t do anything if all conditions return false', function () {
-            const existsStub = sandbox.stub();
-
-            existsStub.withArgs(sinon.match(/config\.js/)).returns(false);
-            existsStub.withArgs(sinon.match(/package\.json/)).returns(false);
-            existsStub.withArgs(sinon.match(/\.ghost-cli/)).returns(true);
-
-            const Command = proxyquire(modulePath, {
-                'fs-extra': {existsSync: existsStub}
-            });
-
-            const exitStub = sandbox.stub(process, 'exit').throws();
-            const errorStub = sandbox.stub(console, 'error');
-
-            Command.checkValidInstall('test');
-            expect(existsStub.calledThrice).to.be.true;
-            expect(errorStub.called).to.be.false;
-            expect(exitStub.called).to.be.false;
         });
     });
 });
