+++ test/unit/utils/check-root-user-spec.js
@@ -4,7 +4,7 @@ const expect = require('chai').expect;
 const sinon = require('sinon');
 const checkRootUser = require('../../../lib/utils/check-root-user');
 
-describe('checkRootUser', function () {
+describe('Unit: Utils > checkRootUser', function () {
     const sandbox = sinon.sandbox.create();
 
     afterEach(() => {
